test: restructure core integration tests by layer and enable SDK e2e in CI
