+++ core/tests/common/fixtures.rs
@@ -1,52 +1,24 @@
+//! User / collateral fixtures and the canonical address normalizer.
+
 use alloy::primitives::{Address, U256, keccak256};
 use alloy::sol_types::SolValue;
 use anyhow::{Result, anyhow};
 use chrono::Utc;
 use core_service::{
-    config::{AppConfig, DEFAULT_ASSET_ADDRESS},
+    config::DEFAULT_ASSET_ADDRESS,
     persist::{PersistCtx, repo},
 };
 use entities::{user, user_asset_balance};
-use migration::{Migrator, MigratorTrait};
 use rand::random;
-use sea_orm::{
-    ColumnTrait, ConnectionTrait, EntityTrait, QueryFilter, Set, Statement, sea_query::OnConflict,
-};
+use sea_orm::{ColumnTrait, EntityTrait, QueryFilter, Set, sea_query::OnConflict};
 use std::str::FromStr;
-use tokio::sync::OnceCell;
-
-static MIGRATIONS: OnceCell<()> = OnceCell::const_new();
-
-pub async fn ensure_migrations(ctx: &PersistCtx) -> Result<()> {
-    MIGRATIONS
-        .get_or_try_init(|| async {
-            Migrator::up(ctx.db.as_ref(), None)
-                .await
-                .map_err(|err| anyhow!("failed to run migrations: {err}"))?;
-            Ok::<(), anyhow::Error>(())
-        })
-        .await?;
-    Ok(())
-}
-
-pub fn init_config() -> Result<AppConfig> {
-    dotenv::dotenv().ok();
-    dotenv::from_filename("../.env").ok();
-    AppConfig::fetch()
-}
-
-pub async fn init_test_env() -> Result<(AppConfig, PersistCtx)> {
-    let cfg = init_config()?;
-    let ctx = PersistCtx::new().await?;
-    ensure_migrations(&ctx).await?;
-    Ok((cfg, ctx))
-}
 
 pub fn random_address() -> String {
     format!("0x{:040x}", random::<u128>())
 }
 
-fn normalize_address(addr: &str) -> Result<String> {
+/// Canonical checksum-free address normalization used across all test layers.
+pub fn normalize_address(addr: &str) -> Result<String> {
     let parsed = Address::from_str(addr).map_err(|err| anyhow!("invalid address {addr}: {err}"))?;
     Ok(format!("{parsed:#x}"))
 }
@@ -91,52 +63,6 @@ pub async fn fetch_user(ctx: &PersistCtx, addr: &str) -> Result<user::Model> {
         .one(ctx.db.as_ref())
         .await?
         .ok_or_else(|| anyhow!("user {addr} not found"))
-}
-
-pub async fn clear_tables(ctx: &PersistCtx, tables: &[&str]) -> Result<()> {
-    for table in tables {
-        ctx.db
-            .as_ref()
-            .execute(Statement::from_string(
-                ctx.db.get_database_backend(),
-                format!(r#"DELETE FROM "{table}";"#),
-            ))
-            .await?;
-    }
-    Ok(())
-}
-
-pub async fn clear_all_tables(ctx: &PersistCtx) -> Result<()> {
-    // Use a single TRUNCATE so FK relationships are handled atomically.
-    ctx.db
-        .as_ref()
-        .execute(Statement::from_string(
-            ctx.db.get_database_backend(),
-            r#"
-TRUNCATE TABLE
-    "UserTransaction",
-    "Withdrawal",
-    "Guarantee",
-    "CycleParticipantPosition",
-    "CycleExposureEdge",
-    "ClearingBatch",
-    "SettlementCycle",
-    "CollateralEvent",
-    "UserAssetBalance",
-    "User",
-    "AuthNonce",
-    "AuthRefreshToken",
-    "WalletRole",
-    "BlockchainEvent",
-    "BlockchainEventCursor",
-    "BlockchainBlock",
-    "ChainCursor"
-RESTART IDENTITY CASCADE;
-"#,
-        ))
-        .await?;
-
-    Ok(())
 }
 
 pub async fn read_user_asset_balance(
