+++ core/tests/common/mod.rs
@@ -1,8 +1,12 @@
 #[allow(dead_code)]
+pub mod api;
+#[allow(dead_code)]
+pub mod chain;
+#[allow(dead_code)]
 pub mod contract;
 #[allow(dead_code)]
 pub mod cycle_fixtures;
 #[allow(dead_code)]
-pub mod fixtures;
+pub mod db;
 #[allow(dead_code)]
-pub mod setup;
+pub mod fixtures;
