+++ core/tests/repo_state.rs
@@ -1,22 +1,131 @@
+//! Repository-layer state machines & idempotency: withdrawals, payment
+//! transactions, cycle exposure/position upserts, wallet-role lookup, and
+//! adversarial inputs. These exercise `repo::` directly against the DB — no
+//! service logic, no chain, no HTTP.
+
 use alloy::primitives::U256;
 use chrono::Utc;
 use core_service::config::DEFAULT_ASSET_ADDRESS;
+use core_service::error::PersistDbError;
 use core_service::persist::repo;
-use entities::sea_orm_active_enums::WithdrawalStatus;
+use entities::sea_orm_active_enums::{
+    ParticipantCycleRole, ParticipantCycleStatus, UserTransactionStatus, WithdrawalStatus,
+};
+use entities::user_transaction;
 use entities::withdrawal::{self, ActiveModel, Entity};
 use sea_orm::{ColumnTrait, EntityTrait, QueryFilter, Set};
 use test_log::test;
 use uuid::Uuid;
 
+#[path = "common/mod.rs"]
 mod common;
-use common::fixtures::{ensure_user, ensure_user_with_collateral, init_test_env, random_address};
 
-use crate::common::fixtures::read_collateral;
+use common::cycle_fixtures::{create_frozen_cycle, setup_cycle_service};
+use common::db::{clear_all_tables, setup_db_test_env};
+use common::fixtures::{ensure_user, ensure_user_with_collateral, random_address, read_collateral};
+use core_service::auth::constants::{SCOPE_GUARANTEE_ISSUE, SCOPE_PAYMENT_READ};
+
+const ADMIN_WALLET_ROLE: &str = "admin";
+const DEFAULT_WALLET_STATUS: &str = "active";
+
+/// Toggle the first lowercase hex digit of an address to upper case, exercising
+/// case-insensitive wallet-role lookup.
+fn mixed_case_address(raw: &str) -> String {
+    let mut toggled = false;
+    let mut out = String::with_capacity(raw.len());
+    for (idx, ch) in raw.chars().enumerate() {
+        if idx < 2 {
+            out.push(ch);
+            continue;
+        }
+        if !toggled && ch.is_ascii_hexdigit() && ch.is_ascii_lowercase() {
+            out.push(ch.to_ascii_uppercase());
+            toggled = true;
+        } else {
+            out.push(ch);
+        }
+    }
+    out
+}
 
 #[test(tokio::test)]
 #[serial_test::file_serial(db)]
+async fn wallet_role_lookup_accepts_mixed_case_wallet_address() -> anyhow::Result<()> {
+    let (_cfg, ctx) = setup_db_test_env().await?;
+    clear_all_tables(&ctx).await?;
+    let signer = alloy::signers::local::PrivateKeySigner::random();
+    let stored_address = format!("{:#x}", signer.address());
+    let presented_address = mixed_case_address(&stored_address);
+
+    assert_eq!(stored_address, presented_address.to_ascii_lowercase());
+    let scopes = vec![
+        SCOPE_GUARANTEE_ISSUE.to_string(),
+        SCOPE_PAYMENT_READ.to_string(),
+    ];
+    repo::upsert_wallet_role(
+        &ctx,
+        &stored_address,
+        ADMIN_WALLET_ROLE,
+        &scopes,
+        DEFAULT_WALLET_STATUS,
+    )
+    .await?;
+    let row = repo::get_wallet_role(&ctx, &presented_address)
+        .await?
+        .expect("wallet role should be retrievable by mixed-case address");
+    assert_eq!(row.role, ADMIN_WALLET_ROLE);
+    assert_eq!(
+        row.scopes,
+        serde_json::json!(["guarantee:issue", "payment:read"])
+    );
+
+    Ok(())
+}
+
+#[test(tokio::test)]
+#[serial_test::file_serial(db)]
+async fn weird_identifiers_do_not_crash() -> anyhow::Result<()> {
+    let (_cfg, ctx) = setup_db_test_env().await?;
+
+    let strange_user = "'; DROP TABLE users; --".to_string();
+    let strange_recipient = "0xdeadbeef::weird".to_string();
+
+    match repo::deposit(
+        &ctx,
+        strange_user.clone(),
+        DEFAULT_ASSET_ADDRESS.to_string(),
+        U256::from(1u64),
+    )
+    .await
+    {
+        Err(PersistDbError::InvariantViolation(_)) => {}
+        Ok(_) => panic!("deposit unexpectedly succeeded for non-existent user"),
+        Err(e) => panic!("unexpected error from deposit: {e}"),
+    }
+
+    match repo::submit_payment_transaction(
+        &ctx,
+        strange_user.clone(),
+        strange_recipient.clone(),
+        DEFAULT_ASSET_ADDRESS.to_string(),
+        "tx::id::odd".into(),
+        U256::from(1u64),
+    )
+    .await
+    {
+        Err(PersistDbError::InvariantViolation(_)) => {}
+        Ok(_) => panic!("submit_payment_transaction unexpectedly succeeded for non-existent user"),
+        Err(e) => panic!("unexpected error from submit_payment_transaction: {e}"),
+    }
+
+    Ok(())
+}
+
+// ════════════════════════ withdrawals (repo state machine) ════════════════════════
+#[test(tokio::test)]
+#[serial_test::file_serial(db)]
 async fn withdrawal_more_than_collateral_fails() -> anyhow::Result<()> {
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
 
     ensure_user_with_collateral(&ctx, &user_addr, U256::from(5u64)).await?;
@@ -42,7 +151,7 @@ async fn withdrawal_more_than_collateral_fails() -> anyhow::Result<()> {
 #[test(tokio::test)]
 #[serial_test::file_serial(db)]
 async fn duplicate_withdrawal_request_updates_existing_pending() -> anyhow::Result<()> {
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
     ensure_user_with_collateral(&ctx, &user_addr, U256::from(20u64)).await?;
 
@@ -107,7 +216,7 @@ async fn duplicate_withdrawal_request_updates_existing_pending() -> anyhow::Resu
 #[test(tokio::test)]
 #[serial_test::file_serial(db)]
 async fn request_withdrawal_after_cancelled_creates_new_pending() -> anyhow::Result<()> {
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
     ensure_user_with_collateral(&ctx, &user_addr, U256::from(20u64)).await?;
 
@@ -167,7 +276,7 @@ async fn request_withdrawal_after_cancelled_creates_new_pending() -> anyhow::Res
 #[test(tokio::test)]
 #[serial_test::file_serial(db)]
 async fn request_withdrawal_after_executed_creates_new_pending() -> anyhow::Result<()> {
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
     ensure_user_with_collateral(&ctx, &user_addr, U256::from(20u64)).await?;
 
@@ -237,7 +346,7 @@ async fn request_withdrawal_after_executed_creates_new_pending() -> anyhow::Resu
 async fn finalize_withdrawal_twice_second_call_errors() -> anyhow::Result<()> {
     use entities::sea_orm_active_enums::WithdrawalStatus;
 
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
 
     ensure_user_with_collateral(&ctx, &user_addr, U256::from(5u64)).await?;
@@ -285,7 +394,7 @@ async fn finalize_withdrawal_twice_second_call_errors() -> anyhow::Result<()> {
 async fn withdrawal_request_cancel_then_finalize_errors() -> anyhow::Result<()> {
     use entities::sea_orm_active_enums::WithdrawalStatus;
 
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
 
     ensure_user_with_collateral(&ctx, &user_addr, U256::from(5u64)).await?;
@@ -342,7 +451,7 @@ async fn withdrawal_request_cancel_then_finalize_errors() -> anyhow::Result<()> 
 #[test(tokio::test)]
 #[serial_test::file_serial(db)]
 async fn finalize_withdrawal_reduces_collateral() -> anyhow::Result<()> {
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
 
     ensure_user_with_collateral(&ctx, &user_addr, U256::from(5u64)).await?;
@@ -373,7 +482,7 @@ async fn finalize_withdrawal_reduces_collateral() -> anyhow::Result<()> {
 #[test(tokio::test)]
 #[serial_test::file_serial(db)]
 async fn finalize_without_any_request_errors_and_preserves_collateral() -> anyhow::Result<()> {
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
 
     ensure_user_with_collateral(&ctx, &user_addr, U256::from(10u64)).await?;
@@ -404,7 +513,7 @@ async fn finalize_without_any_request_errors_and_preserves_collateral() -> anyho
 async fn cancel_after_finalize_does_not_change_executed() -> anyhow::Result<()> {
     use entities::sea_orm_active_enums::WithdrawalStatus;
 
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
 
     ensure_user_with_collateral(&ctx, &user_addr, U256::from(6u64)).await?;
@@ -441,7 +550,7 @@ async fn cancel_after_finalize_does_not_change_executed() -> anyhow::Result<()> 
 async fn double_cancel_is_idempotent() -> anyhow::Result<()> {
     use entities::sea_orm_active_enums::WithdrawalStatus;
 
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
 
     ensure_user_with_collateral(&ctx, &user_addr, U256::from(8u64)).await?;
@@ -469,7 +578,7 @@ async fn double_cancel_is_idempotent() -> anyhow::Result<()> {
 #[test(tokio::test)]
 #[serial_test::file_serial(db)]
 async fn finalize_withdrawal_exceeding_requested_amount_takes_minimum() -> anyhow::Result<()> {
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
 
     ensure_user_with_collateral(&ctx, &user_addr, U256::from(10u64)).await?;
@@ -523,7 +632,7 @@ async fn finalize_withdrawal_records_executed_amount_and_updates_collateral() ->
 {
     use entities::sea_orm_active_enums::WithdrawalStatus;
 
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
 
     // user starts with 10
@@ -581,7 +690,7 @@ async fn finalize_withdrawal_with_full_execution_still_sets_executed_amount() ->
 {
     use entities::sea_orm_active_enums::WithdrawalStatus;
 
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
 
     ensure_user_with_collateral(&ctx, &user_addr, U256::from(10u64)).await?;
@@ -631,7 +740,7 @@ async fn finalize_withdrawal_with_full_execution_still_sets_executed_amount() ->
 #[test(tokio::test)]
 #[serial_test::file_serial(db)]
 async fn unique_pending_withdrawal_per_user_is_enforced() -> anyhow::Result<()> {
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
     ensure_user(&ctx, &user_addr).await?;
 
@@ -700,7 +809,7 @@ async fn unique_pending_withdrawal_per_user_is_enforced() -> anyhow::Result<()> 
 #[test(tokio::test)]
 #[serial_test::file_serial(db)]
 async fn multiple_pending_withdrawals_per_user_different_assets_allowed() -> anyhow::Result<()> {
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
     ensure_user(&ctx, &user_addr).await?;
 
@@ -785,7 +894,7 @@ async fn multiple_pending_withdrawals_per_user_different_assets_allowed() -> any
 async fn deposit_and_withdraw_multiple_assets_updates_collateral_correctly() -> anyhow::Result<()> {
     use entities::sea_orm_active_enums::WithdrawalStatus;
 
-    let (_cfg, ctx) = init_test_env().await?;
+    let (_cfg, ctx) = setup_db_test_env().await?;
     let user_addr = random_address();
 
     // Define two different assets: ETH (default) and a stablecoin
@@ -929,6 +1038,435 @@ async fn deposit_and_withdraw_multiple_assets_updates_collateral_correctly() -> 
         U256::from(50u64).to_string(),
         "Stablecoin executed amount should be 50"
     );
+
+    Ok(())
+}
+
+// ════════════════════════ payment transactions (repo state machine) ════════════════════════
+
+#[serial_test::file_serial(db)]
+async fn duplicate_transaction_id_is_noop() -> anyhow::Result<()> {
+    let (_cfg, ctx) = setup_db_test_env().await?;
+    let user_addr = random_address();
+
+    ensure_user_with_collateral(&ctx, &user_addr, U256::from(5u64)).await?;
+
+    let tx_id = Uuid::new_v4().to_string();
+    let recipient = random_address();
+
+    repo::submit_payment_transaction(
+        &ctx,
+        user_addr.clone(),
+        recipient.clone(),
+        DEFAULT_ASSET_ADDRESS.to_string(),
+        tx_id.clone(),
+        U256::from(2u64),
+    )
+    .await?;
+    repo::submit_payment_transaction(
+        &ctx,
+        user_addr.clone(),
+        recipient,
+        DEFAULT_ASSET_ADDRESS.to_string(),
+        tx_id.clone(),
+        U256::from(2u64),
+    )
+    .await?;
+
+    let txs = user_transaction::Entity::find()
+        .filter(user_transaction::Column::TxId.eq(tx_id))
+        .all(ctx.db.as_ref())
+        .await?;
+    assert_eq!(txs.len(), 1);
+    Ok(())
+}
+
+#[test(tokio::test)]
+#[serial_test::file_serial(db)]
+async fn fail_transaction_twice_is_idempotent() -> anyhow::Result<()> {
+    let (_cfg, ctx) = setup_db_test_env().await?;
+    let user_addr = random_address();
+    let recipient = random_address();
+
+    ensure_user_with_collateral(&ctx, &user_addr, U256::from(10u64)).await?;
+
+    let tx_id = Uuid::new_v4().to_string();
+    repo::submit_payment_transaction(
+        &ctx,
+        user_addr.clone(),
+        recipient,
+        DEFAULT_ASSET_ADDRESS.to_string(),
+        tx_id.clone(),
+        U256::from(3u64),
+    )
+    .await?;
+
+    repo::fail_transaction(&ctx, user_addr.clone(), tx_id.clone()).await?;
+    repo::fail_transaction(&ctx, user_addr.clone(), tx_id.clone()).await?;
+
+    assert_eq!(
+        read_collateral(&ctx, &user_addr, DEFAULT_ASSET_ADDRESS).await?,
+        U256::from(7u64)
+    );
+    Ok(())
+}
+
+// (Dropped `duplicate_tx_id_is_stable_and_idempotent` — exact duplicate of
+// `duplicate_transaction_id_is_noop` above.)
+
+/// Failing a non-existent transaction should error with TransactionNotFound.
+#[test(tokio::test)]
+#[serial_test::file_serial(db)]
+async fn fail_transaction_missing_tx_returns_err() -> anyhow::Result<()> {
+    let (_cfg, ctx) = setup_db_test_env().await?;
+    let user_addr = random_address();
+
+    ensure_user(&ctx, &user_addr).await?;
+
+    let missing_tx_id = Uuid::new_v4().to_string();
+    let res = repo::fail_transaction(&ctx, user_addr.clone(), missing_tx_id.clone()).await;
+
+    match res {
+        Err(PersistDbError::TransactionNotFound(id)) => assert_eq!(id, missing_tx_id),
+        Err(e) => panic!("expected TransactionNotFound, got {e:?}"),
+        Ok(_) => panic!("expected error when tx is missing"),
+    }
+
+    Ok(())
+}
+
+/// NEW: failing a transaction with the wrong user must error and cause no side effects.
+#[test(tokio::test)]
+#[serial_test::file_serial(db)]
+async fn fail_transaction_wrong_user_returns_err_and_no_changes() -> anyhow::Result<()> {
+    let (_cfg, ctx) = setup_db_test_env().await?;
+
+    let owner_addr = random_address();
+    let other_addr = random_address();
+    let recipient = random_address();
+
+    ensure_user_with_collateral(&ctx, &owner_addr, U256::from(10u64)).await?;
+    ensure_user_with_collateral(&ctx, &other_addr, U256::from(10u64)).await?;
+
+    let tx_id = Uuid::new_v4().to_string();
+    repo::submit_payment_transaction(
+        &ctx,
+        owner_addr.clone(),
+        recipient,
+        DEFAULT_ASSET_ADDRESS.to_string(),
+        tx_id.clone(),
+        U256::from(3u64),
+    )
+    .await?;
+
+    // Attempt to fail using the WRONG user address
+    let res = repo::fail_transaction(&ctx, other_addr.clone(), tx_id.clone()).await;
+    assert!(
+        res.is_err(),
+        "expected error when failing tx for the wrong user"
+    );
+
+    // Transaction should remain untouched
+    let row = user_transaction::Entity::find_by_id(tx_id.clone())
+        .one(ctx.db.as_ref())
+        .await?
+        .unwrap();
+    assert!(!row.failed, "tx should not be marked failed");
+    assert!(!row.finalized, "tx should not be finalized");
+
+    // Collateral should be unchanged for both users
+    assert_eq!(
+        read_collateral(&ctx, &owner_addr, DEFAULT_ASSET_ADDRESS).await?,
+        U256::from(10u64)
+    );
+
+    assert_eq!(
+        read_collateral(&ctx, &other_addr, DEFAULT_ASSET_ADDRESS).await?,
+        U256::from(10u64)
+    );
+
+    Ok(())
+}
+
+#[test(tokio::test)]
+#[serial_test::file_serial(db)]
+async fn mark_recorded_accepts_confirmed_status() -> anyhow::Result<()> {
+    let (_cfg, ctx) = setup_db_test_env().await?;
+    let user_addr = random_address();
+    let recipient = random_address();
+    ensure_user_with_collateral(&ctx, &user_addr, U256::from(10u64)).await?;
+
+    let tx_id = Uuid::new_v4().to_string();
+    repo::submit_payment_transaction(
+        &ctx,
+        user_addr.clone(),
+        recipient,
+        DEFAULT_ASSET_ADDRESS.to_string(),
+        tx_id.clone(),
+        U256::from(3u64),
+    )
+    .await?;
+
+    repo::mark_payment_transaction_recorded(
+        &ctx,
+        &tx_id,
+        "0xrecordtx".to_string(),
+        Some(42),
+        Some("0xrecordblock".to_string()),
+    )
+    .await?;
+
+    let row = user_transaction::Entity::find_by_id(tx_id)
+        .one(ctx.db.as_ref())
+        .await?
+        .expect("transaction should exist");
+    assert_eq!(row.status, UserTransactionStatus::Recorded);
+    assert_eq!(row.record_tx_hash.as_deref(), Some("0xrecordtx"));
+    assert_eq!(row.record_tx_block_number, Some(42));
+
+    Ok(())
+}
+
+#[test(tokio::test)]
+#[serial_test::file_serial]
+async fn payment_recording_claim_is_atomic() -> anyhow::Result<()> {
+    let (_cfg, ctx) = setup_db_test_env().await?;
+    let user_addr = random_address();
+    let recipient = random_address();
+    ensure_user(&ctx, &user_addr).await?;
+
+    let tx_id = Uuid::new_v4().to_string();
+    repo::submit_pending_payment_transaction(
+        &ctx,
+        repo::PendingPaymentInput {
+            user_address: user_addr,
+            recipient_address: recipient,
+            asset_address: DEFAULT_ASSET_ADDRESS.to_string(),
+            transaction_id: tx_id.clone(),
+            amount: U256::from(3u64),
+            block_number: 1,
+            block_hash: None,
+        },
+    )
+    .await?;
+
+    assert!(repo::claim_payment_transaction_for_recording(&ctx, &tx_id).await?);
+    assert!(!repo::claim_payment_transaction_for_recording(&ctx, &tx_id).await?);
+
+    let row = user_transaction::Entity::find_by_id(tx_id)
+        .one(ctx.db.as_ref())
+        .await?
+        .expect("transaction should exist");
+    assert_eq!(row.status, UserTransactionStatus::Recording);
+
+    Ok(())
+}
+
+#[test(tokio::test)]
+#[serial_test::file_serial]
+async fn finalized_transaction_cannot_be_marked_reverted() -> anyhow::Result<()> {
+    let (_cfg, ctx) = setup_db_test_env().await?;
+    let user_addr = random_address();
+    let recipient = random_address();
+    ensure_user(&ctx, &user_addr).await?;
+
+    let tx_id = Uuid::new_v4().to_string();
+    repo::submit_pending_payment_transaction(
+        &ctx,
+        repo::PendingPaymentInput {
+            user_address: user_addr,
+            recipient_address: recipient,
+            asset_address: DEFAULT_ASSET_ADDRESS.to_string(),
+            transaction_id: tx_id.clone(),
+            amount: U256::from(3u64),
+            block_number: 1,
+            block_hash: None,
+        },
+    )
+    .await?;
+    repo::mark_payment_transaction_finalized(&ctx, &tx_id).await?;
+
+    assert!(
+        !repo::mark_payment_transaction_reverted(&ctx, &tx_id).await?,
+        "finalized rows must not be reverted by a stale worker"
+    );
+
+    let row = user_transaction::Entity::find_by_id(tx_id)
+        .one(ctx.db.as_ref())
+        .await?
+        .expect("transaction should exist");
+    assert!(row.finalized);
+    assert!(row.verified);
+    assert_eq!(row.status, UserTransactionStatus::Finalized);
+
+    Ok(())
+}
+
+// ════════════════════════ cycle edge/position upserts ════════════════════════
+
+#[serial_test::file_serial(db)]
+async fn replacing_cycle_exposure_edges_is_idempotent() -> anyhow::Result<()> {
+    let service = setup_cycle_service().await?;
+    let ctx = service.persist_ctx();
+    let cycle_id = create_frozen_cycle(ctx, "edge-replace-cycle").await?;
+    let payer = random_address();
+    let payee = random_address();
+
+    let edge = repo::CycleExposureEdgeInput {
+        cycle_id: cycle_id.clone(),
+        payer: payer.clone(),
+        payee: payee.clone(),
+        asset_address: DEFAULT_ASSET_ADDRESS.to_string(),
+        gross_amount: U256::from(5u64),
+        finalized_payable_amount: U256::from(5u64),
+        disputed_amount: U256::ZERO,
+        cancelled_amount: U256::ZERO,
+        guarantee_count: 1,
+    };
+
+    repo::replace_cycle_exposure_edges_on(ctx.db.as_ref(), &cycle_id, vec![edge.clone()]).await?;
+    repo::replace_cycle_exposure_edges_on(ctx.db.as_ref(), &cycle_id, vec![edge]).await?;
+
+    let edges = repo::list_exposure_edges_for_cycle_on(ctx.db.as_ref(), &cycle_id).await?;
+    assert_eq!(edges.len(), 1);
+    assert_eq!(edges[0].gross_amount, "5");
+
+    Ok(())
+}
+
+#[tokio::test]
+#[serial_test::file_serial(db)]
+async fn replacing_participant_positions_removes_stale_rows() -> anyhow::Result<()> {
+    let service = setup_cycle_service().await?;
+    let ctx = service.persist_ctx();
+    let cycle_id = create_frozen_cycle(ctx, "position-replace-cycle").await?;
+    let first = random_address();
+    let second = random_address();
+
+    repo::replace_cycle_participant_positions_on(
+        ctx.db.as_ref(),
+        &cycle_id,
+        vec![
+            repo::CycleParticipantPositionInput {
+                cycle_id: cycle_id.clone(),
+                participant: first,
+                asset_address: DEFAULT_ASSET_ADDRESS.to_string(),
+                gross_outgoing: U256::from(2u64),
+                gross_incoming: U256::ZERO,
+                net_debit: U256::from(2u64),
+                net_credit: U256::ZERO,
+                role: ParticipantCycleRole::NetDebtor,
+                status: ParticipantCycleStatus::Unpaid,
+            },
+            repo::CycleParticipantPositionInput {
+                cycle_id: cycle_id.clone(),
+                participant: second.clone(),
+                asset_address: DEFAULT_ASSET_ADDRESS.to_string(),
+                gross_outgoing: U256::ZERO,
+                gross_incoming: U256::from(2u64),
+                net_debit: U256::ZERO,
+                net_credit: U256::from(2u64),
+                role: ParticipantCycleRole::NetCreditor,
+                status: ParticipantCycleStatus::Claimable,
+            },
+        ],
+    )
+    .await?;
+
+    repo::replace_cycle_participant_positions_on(
+        ctx.db.as_ref(),
+        &cycle_id,
+        vec![repo::CycleParticipantPositionInput {
+            cycle_id: cycle_id.clone(),
+            participant: second,
+            asset_address: DEFAULT_ASSET_ADDRESS.to_string(),
+            gross_outgoing: U256::ZERO,
+            gross_incoming: U256::from(3u64),
+            net_debit: U256::ZERO,
+            net_credit: U256::from(3u64),
+            role: ParticipantCycleRole::NetCreditor,
+            status: ParticipantCycleStatus::Claimable,
+        }],
+    )
+    .await?;
+
+    let positions =
+        repo::list_participant_positions_for_cycle_on(ctx.db.as_ref(), &cycle_id).await?;
+    assert_eq!(positions.len(), 1);
+    assert_eq!(positions[0].net_credit, "3");
+
+    Ok(())
+}
+
+#[tokio::test]
+#[serial_test::file_serial(db)]
+async fn get_user_balance_on_fails_for_nonexistent_user() -> anyhow::Result<()> {
+    let (_cfg, ctx) = setup_db_test_env().await?;
+    let addr = random_address();
+
+    let err = repo::get_user_balance_on(ctx.db.as_ref(), &addr, DEFAULT_ASSET_ADDRESS)
+        .await
+        .expect_err("must fail for unknown user");
+
+    assert!(
+        matches!(err, PersistDbError::UserNotFound(_)),
+        "expected UserNotFound, got: {err:?}"
+    );
+
+    Ok(())
+}
+
+#[tokio::test]
+#[serial_test::file_serial(db)]
+async fn update_user_suspension_increments_version() -> anyhow::Result<()> {
+    let (_cfg, ctx) = setup_db_test_env().await?;
+    let addr = random_address();
+    ensure_user(&ctx, &addr).await?;
+
+    let after_suspend = repo::update_user_suspension(&ctx, &addr, true).await?;
+    assert!(after_suspend.is_suspended);
+    assert_eq!(after_suspend.version, 1);
+
+    let after_unsuspend = repo::update_user_suspension(&ctx, &addr, false).await?;
+    assert!(!after_unsuspend.is_suspended);
+    assert_eq!(after_unsuspend.version, 2);
+
+    Ok(())
+}
+
+#[tokio::test]
+#[serial_test::file_serial(db)]
+async fn store_blockchain_event_duplicate_returns_false() -> anyhow::Result<()> {
+    let (_cfg, ctx) = setup_db_test_env().await?;
+
+    let first = repo::store_blockchain_event(
+        &ctx,
+        1,
+        "Transfer(address,address,uint256)",
+        100,
+        "0xblock",
+        "0xtx",
+        0,
+        "0x0000000000000000000000000000000000000001",
+        "{}",
+    )
+    .await?;
+
+    let second = repo::store_blockchain_event(
+        &ctx,
+        1,
+        "Transfer(address,address,uint256)",
+        100,
+        "0xblock",
+        "0xtx",
+        0,
+        "0x0000000000000000000000000000000000000001",
+        "{}",
+    )
+    .await?;
+
+    assert!(first, "first insert must return true");
+    assert!(!second, "duplicate insert must return false");
 
     Ok(())
 }
