+++ core/tests/service_settlement.rs
@@ -1,3 +1,7 @@
+//! Service-layer settlement: participant status transitions, default handling,
+//! and recipient payment queryability. No chain, no HTTP.
+
+use alloy::primitives::U256;
 use chrono::Utc;
 use entities::sea_orm_active_enums::{ParticipantCycleStatus, SettlementCycleStatus};
 use entities::settlement_cycle;
@@ -7,7 +11,8 @@ use sea_orm::EntityTrait;
 mod common;
 
 use common::cycle_fixtures::{build_three_party_cycle, setup_cycle_service};
-use core_service::persist::repo;
+use common::fixtures::{ensure_user_with_collateral, random_address};
+use core_service::{config::DEFAULT_ASSET_ADDRESS, persist::repo};
 
 #[tokio::test]
 #[serial_test::file_serial(db)]
@@ -106,6 +111,33 @@ async fn unpaid_debtors_move_payment_window_cycle_to_defaulted() -> anyhow::Resu
         .await?
         .expect("cycle exists");
     assert_eq!(cycle.status, SettlementCycleStatus::Defaulted);
+
+    Ok(())
+}
+
+#[tokio::test]
+#[serial_test::file_serial(db)]
+async fn recipient_payments_are_queryable_by_recipient() -> anyhow::Result<()> {
+    let service = setup_cycle_service().await?;
+    let ctx = service.persist_ctx();
+    let user = random_address();
+    let recipient = random_address();
+    ensure_user_with_collateral(ctx, &user, U256::from(25u64)).await?;
+
+    repo::submit_payment_transaction(
+        ctx,
+        user.clone(),
+        recipient.clone(),
+        DEFAULT_ASSET_ADDRESS.to_string(),
+        "0xpayment".to_string(),
+        U256::from(9u64),
+    )
+    .await?;
+
+    let rows = repo::get_recipient_transactions(ctx, &recipient).await?;
+    assert_eq!(rows.len(), 1);
+    assert_eq!(rows[0].user_address, user);
+    assert_eq!(rows[0].amount, "9");
 
     Ok(())
 }
